A 校友圈接口
