+++ src/main/java/cn/edu/seu/alumni_server/controller/AlumniCircleController.java
@@ -6,14 +6,19 @@ import cn.edu.seu.alumni_server.common.exceptions.AlumniCircleServiceException;
 import cn.edu.seu.alumni_server.controller.dto.MyAlumniCircleInfoDTO;
 import cn.edu.seu.alumni_server.controller.dto.PageResult;
 import cn.edu.seu.alumni_server.controller.dto.alumnicircle.AlumniCircleBasicInfoDTO;
+import cn.edu.seu.alumni_server.controller.dto.alumnicircle.AlumniCircleDTO;
+import cn.edu.seu.alumni_server.controller.dto.alumnicircle.AlumniCircleMemberDTO;
+import cn.edu.seu.alumni_server.dao.entity.Activity;
+import cn.edu.seu.alumni_server.dao.entity.AlumniCircleMember;
+import cn.edu.seu.alumni_server.dao.mapper.ActivityMapper;
+import cn.edu.seu.alumni_server.dao.mapper.AlumniCircleMapper;
+import cn.edu.seu.alumni_server.dao.mapper.AlumniCircleMemberMapper;
 import cn.edu.seu.alumni_server.service.AlumniCircleService;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+import tk.mybatis.mapper.entity.Example;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.LinkedList;
@@ -22,6 +27,7 @@ import java.util.List;
 @SuppressWarnings("ALL")
 @RestController
 //@Acl
+@RequestMapping("/alumniCircle")
 public class AlumniCircleController {
 
     @Autowired
@@ -30,7 +36,17 @@ public class AlumniCircleController {
     @Autowired
     AlumniCircleService alumniCircleService;
 
-    @GetMapping("/alumniCircles/enrolledAlumniCircles")
+    @Autowired
+    ActivityMapper activityMapper;
+
+    @Autowired
+    AlumniCircleMapper alumniCircleMapper;
+
+
+    @Autowired
+    AlumniCircleMemberMapper alumniCircleMemberMapper;
+
+    @GetMapping("/enrolledAlumniCircles")
     public WebResponse getEnrolledAlumniCirclesByAccountId(
             HttpServletRequest request,
             @RequestParam(value = "accountId", required = false)
@@ -58,7 +74,7 @@ public class AlumniCircleController {
         }
     }
 
-    @GetMapping("/search/alumniCircles")
+    @GetMapping("/search")
     public WebResponse searchByActivityName(
             @RequestParam(value = "alumniCircleName", required = true) String alumniCircleName,
             @RequestParam(value = "fuzzy", required = false, defaultValue = "true") Boolean fuzzy,
@@ -85,9 +101,84 @@ public class AlumniCircleController {
         }
     }
 
-    @RequestMapping("/alumniCircles/recommend")
+    @RequestMapping("/recommend")
     public WebResponse alumniCirclesRecommend() {
         return new WebResponse().success(
                 alumniCircleService.alumniCirclesRecommend());
     }
+
+    @GetMapping("/information")
+    public WebResponse information(@RequestParam Long alumniCircleId) {
+        alumniCircleMapper.selectByPrimaryKey(alumniCircleId);
+        return new WebResponse().success(
+                alumniCircleMapper.selectByPrimaryKey(alumniCircleId));
+    }
+
+
+    @PostMapping("/join")
+    public WebResponse join(@RequestParam Long alumniCircleId,
+                            @RequestParam Long accountId) {
+
+        AlumniCircleMember alumniCircleMember = new AlumniCircleMember();
+        alumniCircleMember.setAccountId(accountId);
+        alumniCircleMember.setAlumniCircleId(alumniCircleId);
+        alumniCircleMemberMapper.insert(alumniCircleMember);
+
+        return new WebResponse();
+    }
+
+    @PostMapping("/leave")
+    public WebResponse leave(@RequestParam Long alumniCircleId,
+                             @RequestParam Long accountId) {
+
+        AlumniCircleMember alumniCircleMember = new AlumniCircleMember();
+        alumniCircleMember.setAccountId(accountId);
+        alumniCircleMember.setAlumniCircleId(alumniCircleId);
+        alumniCircleMember.setValidStatus(false);
+        alumniCircleMemberMapper.insert(alumniCircleMember);
+        return new WebResponse();
+    }
+
+    @GetMapping("/members")
+    public WebResponse members(
+            @RequestParam Long alumniCircleId,
+            @RequestParam int pageIndex,
+            @RequestParam int pageSize
+    ) {
+        PageHelper.startPage(pageIndex, pageSize);
+
+        Example example = new Example(AlumniCircleMember.class);
+        example.createCriteria().andEqualTo("alumniCircleId", alumniCircleId);
+        List<AlumniCircleMemberDTO> res = alumniCircleMemberMapper.getAlumniCircleMembers();
+
+        return new WebResponse().success(
+                new PageResult<AlumniCircleMemberDTO>(((Page) res).getTotal(), res));
+    }
+
+    /**
+     * 活动列表
+     *
+     * @return
+     */
+    @PostMapping("/activities")
+    public WebResponse activities(@RequestParam Long alumniCircleId,
+                                  @RequestParam int pageIndex,
+                                  @RequestParam int pageSize
+    ) {
+        Activity activity = new Activity();
+        activity.setAlumniCircleId(alumniCircleId);
+
+        PageHelper.startPage(pageIndex, pageSize);
+        List<Activity> activities = activityMapper.select(activity);
+
+        return new WebResponse();
+    }
+
+    @PutMapping("/maintain")
+    public WebResponse maintain(@RequestBody AlumniCircleDTO alumniCircleDTO) {
+
+        alumniCircleMapper.updateByPrimaryKey(alumniCircleDTO.toAlumniCircle());
+        return new WebResponse();
+    }
+
 }
